Add BaseView and finalize classes View and WebView
