--- tests/Mocks/WebViewPlaceholderMock.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\View\Tests\Mocks;
-
-use Yiisoft\View\WebView;
-
-class WebViewPlaceholderMock extends WebView
-{
-    public function endPage($ajaxMode = false): void
-    {
-        $this->setPlaceholderSalt((string)time());
-        parent::endPage($ajaxMode);
-    }
-}
